+++ lib/src/store.rs
@@ -1,6 +1,7 @@
 //! In-memory store of Atomic data.
 //! This provides many methods for finding, changing, serializing and parsing Atomic Data.
 
+use crate::agents::Agent;
 use crate::storelike::QueryResult;
 use crate::Value;
 use crate::{atoms::Atom, storelike::Storelike};
@@ -8,6 +9,7 @@ use crate::{errors::AtomicResult, Resource};
 use std::{collections::HashMap, sync::Arc, sync::Mutex};
 
 /// The in-memory store of data, containing the Resources, Properties and Classes
+/// It uses the `default_agent` as the default client.
 #[derive(Clone)]
 pub struct Store {
     // The store currently holds two stores - that is not ideal
@@ -166,7 +168,7 @@ impl Storelike for Store {
         Some(self.get_server_url().into())
     }
 
-    fn get_default_agent(&self) -> AtomicResult<crate::agents::Agent> {
+    fn get_default_agent(&self) -> AtomicResult<Agent> {
         match self.default_agent.lock().unwrap().to_owned() {
             Some(agent) => Ok(agent),
             None => Err("No default agent has been set.".into()),
@@ -177,7 +179,11 @@ impl Storelike for Store {
         if let Some(resource) = self.hashmap.lock().unwrap().get(subject) {
             return Ok(resource.clone());
         }
-        self.handle_not_found(subject, "Not found in HashMap.".into())
+        self.handle_not_found(
+            subject,
+            "Not found in HashMap.".into(),
+            self.get_default_agent().ok().as_ref(),
+        )
     }
 
     fn remove_resource(&self, subject: &str) -> AtomicResult<()> {
@@ -192,7 +198,7 @@ impl Storelike for Store {
         Ok(())
     }
 
-    fn set_default_agent(&self, agent: crate::agents::Agent) {
+    fn set_default_agent(&self, agent: Agent) {
         self.default_agent.lock().unwrap().replace(agent);
     }