+++ CHANGELOG.md
@@ -14,6 +14,7 @@ See [STATUS.md](server/STATUS.md) to learn more about which features will remain
 - Add `clear` option to error component (resets all front-end state)
 - Add `Agent::from_secret` #785
 - Make `set_propval` and `set_propval_shortname` chainable #785
+- Don't use default agent when fetching with Db #787
 
 ## [v0.36.1] - 2023-12-06